+++ openfisca_core/periods/helpers.py
@@ -1,26 +1,29 @@
-from typing import NoReturn, Optional
+from __future__ import annotations
+
+from typing import NoReturn
 
 import datetime
-import os
+import functools
 
 import pendulum
-from pendulum.parsing import ParserError
 
-from . import _parsers, config
+from . import config, types as t
+from ._errors import InstantError, PeriodError
+from ._parsers import parse_instant, parse_period
 from .date_unit import DateUnit
 from .instant_ import Instant
 from .period_ import Period
 
 
-def instant(instant) -> Optional[Instant]:
+@functools.singledispatch
+def instant(value: object) -> t.Instant:
     """Build a new instant, aka a triple of integers (year, month, day).
 
     Args:
-        instant: An ``instant-like`` object.
+        value(object): An ``instant-like`` object.
 
     Returns:
-        None: When ``instant`` is None.
-        :obj:`.Instant`: Otherwise.
+        :obj:`.Instant`: A new instant.
 
     Raises:
         :exc:`ValueError`: When the arguments were invalid, like "2021-32-13".
@@ -47,38 +50,55 @@ def instant(instant) -> Optional[Instant]:
         >>> instant("2021")
         Instant((2021, 1, 1))
 
+        >>> instant([2021])
+        Instant((2021, 1, 1))
+
+        >>> instant([2021, 9])
+        Instant((2021, 9, 1))
+
+        >>> instant(None)
+        Traceback (most recent call last):
+        openfisca_core.periods._errors.InstantError: 'None' is not a valid i...
+
     """
-    if instant is None:
-        return None
-    if isinstance(instant, Instant):
-        return instant
-    if isinstance(instant, str):
-        if not config.INSTANT_PATTERN.match(instant):
-            msg = f"'{instant}' is not a valid instant. Instants are described using the 'YYYY-MM-DD' format, for instance '2015-06-15'."
-            raise ValueError(
-                msg,
-            )
-        instant = Instant(int(fragment) for fragment in instant.split("-", 2)[:3])
-    elif isinstance(instant, datetime.date):
-        instant = Instant((instant.year, instant.month, instant.day))
-    elif isinstance(instant, int):
-        instant = (instant,)
-    elif isinstance(instant, list):
-        assert 1 <= len(instant) <= 3
-        instant = tuple(instant)
-    elif isinstance(instant, Period):
-        instant = instant.start
-    else:
-        assert isinstance(instant, tuple), instant
-        assert 1 <= len(instant) <= 3
-    if len(instant) == 1:
-        return Instant((instant[0], 1, 1))
-    if len(instant) == 2:
-        return Instant((instant[0], instant[1], 1))
-    return Instant(instant)
+
+    if isinstance(value, t.SeqInt):
+        return Instant((list(value) + [1] * 3)[:3])
+
+    raise InstantError(str(value))
 
 
-def instant_date(instant: Optional[Instant]) -> Optional[datetime.date]:
+@instant.register
+def _(value: None) -> NoReturn:
+    raise InstantError(str(value))
+
+
+@instant.register
+def _(value: int) -> t.Instant:
+    return Instant((value, 1, 1))
+
+
+@instant.register
+def _(value: Period) -> t.Instant:
+    return value.start
+
+
+@instant.register
+def _(value: t.Instant) -> t.Instant:
+    return value
+
+
+@instant.register
+def _(value: datetime.date) -> t.Instant:
+    return Instant((value.year, value.month, value.day))
+
+
+@instant.register
+def _(value: str) -> t.Instant:
+    return parse_instant(value)
+
+
+def instant_date(instant: None | t.Instant) -> None | datetime.date:
     """Returns the date representation of an :class:`.Instant`.
 
     Args:
@@ -104,7 +124,8 @@ def instant_date(instant: Optional[Instant]) -> Optional[datetime.date]:
     return instant_date
 
 
-def period(value) -> Period:
+@functools.singledispatch
+def period(value: object) -> t.Period:
     """Build a new period, aka a triple (unit, start_instant, size).
 
     Args:
@@ -124,7 +145,7 @@ def period(value) -> Period:
         Period((<DateUnit.DAY: 'day'>, Instant((2021, 1, 1)), 1))
 
         >>> period(DateUnit.ETERNITY)
-        Period((<DateUnit.ETERNITY: 'eternity'>, Instant((1, 1, 1)), inf))
+        Period((<DateUnit.ETERNITY: 'eternity'>, Instant((-1, -1, -1)), -1))
 
         >>> period(2021)
         Period((<DateUnit.YEAR: 'year'>, Instant((2021, 1, 1)), 1))
@@ -147,124 +168,86 @@ def period(value) -> Period:
         >>> period("day:2014-02-02:3")
         Period((<DateUnit.DAY: 'day'>, Instant((2014, 2, 2)), 3))
 
-
     """
-    if isinstance(value, Period):
-        return value
 
-    # We return a "day-period", for example
-    # ``<Period(('day', <Instant(2021, 1, 1)>, 1))>``.
-    if isinstance(value, Instant):
-        return Period((DateUnit.DAY, value, 1))
-
-    # For example ``datetime.date(2021, 9, 16)``.
-    if isinstance(value, datetime.date):
-        return Period((DateUnit.DAY, instant(value), 1))
+    one, two, three = 1, 2, 3
 
     # We return an "eternity-period", for example
-    # ``<Period(('eternity', <Instant(1, 1, 1)>, inf))>``.
+    # ``<Period(('eternity', <Instant(-1, -1, -1)>, -1))>``.
     if str(value).lower() == DateUnit.ETERNITY:
-        return Period(
-            (
-                DateUnit.ETERNITY,
-                instant(datetime.date.min),
-                float("inf"),
-            ),
-        )
+        return Period.eternity()
 
-    # For example ``2021`` gives
-    # ``<Period(('year', <Instant(2021, 1, 1)>, 1))>``.
-    if isinstance(value, int):
-        return Period((DateUnit.YEAR, instant(value), 1))
-
-    # Up to this point, if ``value`` is not a :obj:`str`, we desist.
-    if not isinstance(value, str):
-        _raise_error(value)
-
-    # There can't be empty strings.
-    if not value:
-        _raise_error(value)
-
-    # Try to parse from an ISO format/calendar period.
-    try:
-        period = _parsers._parse_period(value)
-
-    except (AttributeError, ParserError, ValueError):
-        _raise_error(value)
-
-    if period is not None:
-        return period
+    # We try to parse from an ISO format/calendar period.
+    if isinstance(value, t.InstantStr):
+        return parse_period(value)
 
     # A complex period has a ':' in its string.
-    if ":" not in value:
-        _raise_error(value)
+    if isinstance(value, t.PeriodStr):
+        components = value.split(":")
 
-    components = value.split(":")
+        # The left-most component must be a valid unit
+        unit = components[0]
 
-    # left-most component must be a valid unit
-    unit = components[0]
+        if unit not in list(DateUnit) or unit == DateUnit.ETERNITY:
+            raise PeriodError(str(value))
 
-    if unit not in list(DateUnit) or unit == DateUnit.ETERNITY:
-        _raise_error(value)
+        # Cast ``unit`` to DateUnit.
+        unit = DateUnit(unit)
 
-    # Cast ``unit`` to DateUnit.
-    unit = DateUnit(unit)
+        # The middle component must be a valid iso period
+        period = parse_period(components[1])
 
-    # middle component must be a valid iso period
-    try:
-        base_period = _parsers._parse_period(components[1])
+        # Periods like year:2015-03 have a size of 1
+        if len(components) == two:
+            size = one
 
-    except (AttributeError, ParserError, ValueError):
-        _raise_error(value)
+        # if provided, make sure the size is an integer
+        elif len(components) == three:
+            try:
+                size = int(components[2])
 
-    if not base_period:
-        _raise_error(value)
+            except ValueError as error:
+                raise PeriodError(str(value)) from error
 
-    # period like year:2015-03 have a size of 1
-    if len(components) == 2:
-        size = 1
+        # If there are more than 2 ":" in the string, the period is invalid
+        else:
+            raise PeriodError(str(value))
 
-    # if provided, make sure the size is an integer
-    elif len(components) == 3:
-        try:
-            size = int(components[2])
+        # Reject ambiguous periods such as month:2014
+        if unit_weight(period.unit) > unit_weight(unit):
+            raise PeriodError(str(value))
 
-        except ValueError:
-            _raise_error(value)
+        return Period((unit, period.start, size))
 
-    # if there is more than 2 ":" in the string, the period is invalid
-    else:
-        _raise_error(value)
-
-    # reject ambiguous periods such as month:2014
-    if unit_weight(base_period.unit) > unit_weight(unit):
-        _raise_error(value)
-
-    return Period((unit, base_period.start, size))
+    raise PeriodError(str(value))
 
 
-def _raise_error(value: str) -> NoReturn:
-    """Raise an error.
-
-    Examples:
-        >>> _raise_error("Oi mate!")
-        Traceback (most recent call last):
-        ValueError: Expected a period (eg. '2017', '2017-01', '2017-01-01', ...
-        Learn more about legal period formats in OpenFisca:
-        <https://openfisca.org/doc/coding-the-legislation/35_periods.html#pe...
-
-    """
-    message = os.linesep.join(
-        [
-            f"Expected a period (eg. '2017', '2017-01', '2017-01-01', ...); got: '{value}'.",
-            "Learn more about legal period formats in OpenFisca:",
-            "<https://openfisca.org/doc/coding-the-legislation/35_periods.html#periods-in-simulations>.",
-        ],
-    )
-    raise ValueError(message)
+@period.register
+def _(value: None) -> NoReturn:
+    raise PeriodError(str(value))
 
 
-def key_period_size(period: Period) -> str:
+@period.register
+def _(value: int) -> t.Period:
+    return Period((DateUnit.YEAR, instant(value), 1))
+
+
+@period.register
+def _(value: t.Period) -> t.Period:
+    return value
+
+
+@period.register
+def _(value: t.Instant) -> t.Period:
+    return Period((DateUnit.DAY, value, 1))
+
+
+@period.register
+def _(value: datetime.date) -> t.Period:
+    return Period((DateUnit.DAY, instant(value), 1))
+
+
+def key_period_size(period: t.Period) -> str:
     """Define a key in order to sort periods by length.
 
     It uses two aspects: first, ``unit``, then, ``size``.
@@ -287,12 +270,11 @@ def key_period_size(period: Period) -> str:
         '300_3'
 
     """
-    unit, start, size = period
 
-    return f"{unit_weight(unit)}_{size}"
+    return f"{unit_weight(period.unit)}_{period.size}"
 
 
-def unit_weights() -> dict[str, int]:
+def unit_weights() -> dict[t.DateUnit, int]:
     """Assign weights to date units.
 
     Examples:
@@ -310,7 +292,7 @@ def unit_weights() -> dict[str, int]:
     }
 
 
-def unit_weight(unit: str) -> int:
+def unit_weight(unit: t.DateUnit) -> int:
     """Retrieves a specific date unit weight.
 
     Examples:
@@ -319,3 +301,13 @@ def unit_weight(unit: str) -> int:
 
     """
     return unit_weights()[unit]
+
+
+__all__ = [
+    "instant",
+    "instant_date",
+    "key_period_size",
+    "period",
+    "unit_weight",
+    "unit_weights",
+]