+++ openfisca_core/periods/period_.py
@@ -1,23 +1,18 @@
 from __future__ import annotations
 
-import typing
+from collections.abc import Sequence
 
 import calendar
 import datetime
 
 import pendulum
 
-from . import helpers
+from . import helpers, types as t
 from .date_unit import DateUnit
 from .instant_ import Instant
 
-if typing.TYPE_CHECKING:
-    from collections.abc import Sequence
 
-    from pendulum.datetime import Date
-
-
-class Period(tuple):
+class Period(tuple[t.DateUnit, t.Instant, int]):
     """Toolbox to handle date intervals.
 
     A :class:`.Period` is a triple (``unit``, ``start``, ``size``).
@@ -122,14 +117,16 @@ class Period(tuple):
 
     """
 
+    __slots__ = ()
+
     def __repr__(self) -> str:
         return f"{self.__class__.__name__}({super().__repr__()})"
 
-    def __str__(self) -> str:
+    def __str__(self) -> t.PeriodStr:
         unit, start_instant, size = self
 
         if unit == DateUnit.ETERNITY:
-            return unit.upper()
+            return t.PeriodStr(unit.upper())
 
         # ISO format date units.
         f_year, month, day = start_instant
@@ -141,56 +138,56 @@ class Period(tuple):
         if unit == DateUnit.MONTH and size == 12 or unit == DateUnit.YEAR and size == 1:
             if month == 1:
                 # civil year starting from january
-                return str(f_year)
+                return t.PeriodStr(str(f_year))
             # rolling year
-            return f"{DateUnit.YEAR}:{f_year}-{month:02d}"
+            return t.PeriodStr(f"{DateUnit.YEAR}:{f_year}-{month:02d}")
 
         # simple month
         if unit == DateUnit.MONTH and size == 1:
-            return f"{f_year}-{month:02d}"
+            return t.PeriodStr(f"{f_year}-{month:02d}")
 
         # several civil years
         if unit == DateUnit.YEAR and month == 1:
-            return f"{unit}:{f_year}:{size}"
+            return t.PeriodStr(f"{unit}:{f_year}:{size}")
 
         if unit == DateUnit.DAY:
             if size == 1:
-                return f"{f_year}-{month:02d}-{day:02d}"
-            return f"{unit}:{f_year}-{month:02d}-{day:02d}:{size}"
+                return t.PeriodStr(f"{f_year}-{month:02d}-{day:02d}")
+            return t.PeriodStr(f"{unit}:{f_year}-{month:02d}-{day:02d}:{size}")
 
         # 1 week
         if unit == DateUnit.WEEK and size == 1:
             if week < 10:
-                return f"{c_year}-W0{week}"
+                return t.PeriodStr(f"{c_year}-W0{week}")
 
-            return f"{c_year}-W{week}"
+            return t.PeriodStr(f"{c_year}-W{week}")
 
         # several weeks
         if unit == DateUnit.WEEK and size > 1:
             if week < 10:
-                return f"{unit}:{c_year}-W0{week}:{size}"
+                return t.PeriodStr(f"{unit}:{c_year}-W0{week}:{size}")
 
-            return f"{unit}:{c_year}-W{week}:{size}"
+            return t.PeriodStr(f"{unit}:{c_year}-W{week}:{size}")
 
         # 1 weekday
         if unit == DateUnit.WEEKDAY and size == 1:
             if week < 10:
-                return f"{c_year}-W0{week}-{weekday}"
+                return t.PeriodStr(f"{c_year}-W0{week}-{weekday}")
 
-            return f"{c_year}-W{week}-{weekday}"
+            return t.PeriodStr(f"{c_year}-W{week}-{weekday}")
 
         # several weekdays
         if unit == DateUnit.WEEKDAY and size > 1:
             if week < 10:
-                return f"{unit}:{c_year}-W0{week}-{weekday}:{size}"
+                return t.PeriodStr(f"{unit}:{c_year}-W0{week}-{weekday}:{size}")
 
-            return f"{unit}:{c_year}-W{week}-{weekday}:{size}"
+            return t.PeriodStr(f"{unit}:{c_year}-W{week}-{weekday}:{size}")
 
         # complex period
-        return f"{unit}:{f_year}-{month:02d}:{size}"
+        return t.PeriodStr(f"{unit}:{f_year}-{month:02d}:{size}")
 
     @property
-    def unit(self) -> str:
+    def unit(self) -> t.DateUnit:
         """The ``unit`` of the ``Period``.
 
         Example:
@@ -203,7 +200,7 @@ class Period(tuple):
         return self[0]
 
     @property
-    def start(self) -> Instant:
+    def start(self) -> t.Instant:
         """The ``Instant`` at which the ``Period`` starts.
 
         Example:
@@ -229,7 +226,7 @@ class Period(tuple):
         return self[2]
 
     @property
-    def date(self) -> Date:
+    def date(self) -> pendulum.Date:
         """The date representation of the ``Period`` start date.
 
         Examples:
@@ -317,7 +314,12 @@ class Period(tuple):
 
         """
         if self.unit in (DateUnit.YEAR, DateUnit.MONTH):
-            last_day = self.start.offset(self.size, self.unit).offset(-1, DateUnit.DAY)
+            last = self.start.offset(self.size, self.unit)
+            if last is None:
+                raise NotImplementedError
+            last_day = last.offset(-1, DateUnit.DAY)
+            if last_day is None:
+                raise NotImplementedError
             return (last_day.date - self.start.date).days + 1
 
         if self.unit == DateUnit.WEEK:
@@ -330,7 +332,7 @@ class Period(tuple):
         raise ValueError(msg)
 
     @property
-    def size_in_weeks(self):
+    def size_in_weeks(self) -> int:
         """The ``size`` of the ``Period`` in weeks.
 
         Examples:
@@ -348,14 +350,14 @@ class Period(tuple):
         if self.unit == DateUnit.YEAR:
             start = self.start.date
             cease = start.add(years=self.size)
-            delta = pendulum.period(start, cease)
-            return delta.as_interval().weeks
+            delta = start.diff(cease)
+            return delta.in_weeks()
 
         if self.unit == DateUnit.MONTH:
             start = self.start.date
             cease = start.add(months=self.size)
-            delta = pendulum.period(start, cease)
-            return delta.as_interval().weeks
+            delta = start.diff(cease)
+            return delta.in_weeks()
 
         if self.unit == DateUnit.WEEK:
             return self.size
@@ -364,7 +366,7 @@ class Period(tuple):
         raise ValueError(msg)
 
     @property
-    def size_in_weekdays(self):
+    def size_in_weekdays(self) -> int:
         """The ``size`` of the ``Period`` in weekdays.
 
         Examples:
@@ -382,8 +384,13 @@ class Period(tuple):
         if self.unit == DateUnit.YEAR:
             return self.size_in_weeks * 7
 
-        if self.unit in DateUnit.MONTH:
-            last_day = self.start.offset(self.size, self.unit).offset(-1, DateUnit.DAY)
+        if DateUnit.MONTH in self.unit:
+            last = self.start.offset(self.size, self.unit)
+            if last is None:
+                raise NotImplementedError
+            last_day = last.offset(-1, DateUnit.DAY)
+            if last_day is None:
+                raise NotImplementedError
             return (last_day.date - self.start.date).days + 1
 
         if self.unit == DateUnit.WEEK:
@@ -396,11 +403,13 @@ class Period(tuple):
         raise ValueError(msg)
 
     @property
-    def days(self):
+    def days(self) -> int:
         """Same as ``size_in_days``."""
         return (self.stop.date - self.start.date).days + 1
 
-    def intersection(self, start, stop):
+    def intersection(
+        self, start: t.Instant | None, stop: t.Instant | None
+    ) -> t.Period | None:
         if start is None and stop is None:
             return self
         period_start = self[1]
@@ -453,17 +462,17 @@ class Period(tuple):
             ),
         )
 
-    def get_subperiods(self, unit: DateUnit) -> Sequence[Period]:
+    def get_subperiods(self, unit: t.DateUnit) -> Sequence[t.Period]:
         """Return the list of periods of unit ``unit`` contained in self.
 
         Examples:
             >>> period = Period((DateUnit.YEAR, Instant((2021, 1, 1)), 1))
             >>> period.get_subperiods(DateUnit.MONTH)
-            [Period((<DateUnit.MONTH: 'month'>, Instant((2021, 1, 1)), 1)),...2021, 12, 1)), 1))]
+            [Period((<DateUnit.MONTH: 'month'>, Instant((2021, 1, 1)), 1)),...]
 
             >>> period = Period((DateUnit.YEAR, Instant((2021, 1, 1)), 2))
             >>> period.get_subperiods(DateUnit.YEAR)
-            [Period((<DateUnit.YEAR: 'year'>, Instant((2021, 1, 1)), 1)),...((2022, 1, 1)), 1))]
+            [Period((<DateUnit.YEAR: 'year'>, Instant((2021, 1, 1)), 1)), P...]
 
         """
         if helpers.unit_weight(self.unit) < helpers.unit_weight(unit):
@@ -499,26 +508,34 @@ class Period(tuple):
         msg = f"Cannot subdivide {self.unit} into {unit}"
         raise ValueError(msg)
 
-    def offset(self, offset, unit=None):
+    def offset(self, offset: str | int, unit: t.DateUnit | None = None) -> t.Period:
         """Increment (or decrement) the given period with offset units.
 
         Examples:
             >>> Period((DateUnit.DAY, Instant((2021, 1, 1)), 365)).offset(1)
             Period((<DateUnit.DAY: 'day'>, Instant((2021, 1, 2)), 365))
 
-            >>> Period((DateUnit.DAY, Instant((2021, 1, 1)), 365)).offset(1, DateUnit.DAY)
+            >>> Period((DateUnit.DAY, Instant((2021, 1, 1)), 365)).offset(
+            ...     1, DateUnit.DAY
+            ... )
             Period((<DateUnit.DAY: 'day'>, Instant((2021, 1, 2)), 365))
 
-            >>> Period((DateUnit.DAY, Instant((2021, 1, 1)), 365)).offset(1, DateUnit.MONTH)
+            >>> Period((DateUnit.DAY, Instant((2021, 1, 1)), 365)).offset(
+            ...     1, DateUnit.MONTH
+            ... )
             Period((<DateUnit.DAY: 'day'>, Instant((2021, 2, 1)), 365))
 
-            >>> Period((DateUnit.DAY, Instant((2021, 1, 1)), 365)).offset(1, DateUnit.YEAR)
+            >>> Period((DateUnit.DAY, Instant((2021, 1, 1)), 365)).offset(
+            ...     1, DateUnit.YEAR
+            ... )
             Period((<DateUnit.DAY: 'day'>, Instant((2022, 1, 1)), 365))
 
             >>> Period((DateUnit.MONTH, Instant((2021, 1, 1)), 12)).offset(1)
             Period((<DateUnit.MONTH: 'month'>, Instant((2021, 2, 1)), 12))
 
-            >>> Period((DateUnit.MONTH, Instant((2021, 1, 1)), 12)).offset(1, DateUnit.DAY)
+            >>> Period((DateUnit.MONTH, Instant((2021, 1, 1)), 12)).offset(
+            ...     1, DateUnit.DAY
+            ... )
             Period((<DateUnit.MONTH: 'month'>, Instant((2021, 1, 2)), 12))
 
             >>> Period((DateUnit.MONTH, Instant((2021, 1, 1)), 12)).offset(
@@ -526,19 +543,27 @@ class Period(tuple):
             ... )
             Period((<DateUnit.MONTH: 'month'>, Instant((2021, 2, 1)), 12))
 
-            >>> Period((DateUnit.MONTH, Instant((2021, 1, 1)), 12)).offset(1, DateUnit.YEAR)
+            >>> Period((DateUnit.MONTH, Instant((2021, 1, 1)), 12)).offset(
+            ...     1, DateUnit.YEAR
+            ... )
             Period((<DateUnit.MONTH: 'month'>, Instant((2022, 1, 1)), 12))
 
             >>> Period((DateUnit.YEAR, Instant((2021, 1, 1)), 1)).offset(1)
             Period((<DateUnit.YEAR: 'year'>, Instant((2022, 1, 1)), 1))
 
-            >>> Period((DateUnit.YEAR, Instant((2021, 1, 1)), 1)).offset(1, DateUnit.DAY)
+            >>> Period((DateUnit.YEAR, Instant((2021, 1, 1)), 1)).offset(
+            ...     1, DateUnit.DAY
+            ... )
             Period((<DateUnit.YEAR: 'year'>, Instant((2021, 1, 2)), 1))
 
-            >>> Period((DateUnit.YEAR, Instant((2021, 1, 1)), 1)).offset(1, DateUnit.MONTH)
+            >>> Period((DateUnit.YEAR, Instant((2021, 1, 1)), 1)).offset(
+            ...     1, DateUnit.MONTH
+            ... )
             Period((<DateUnit.YEAR: 'year'>, Instant((2021, 2, 1)), 1))
 
-            >>> Period((DateUnit.YEAR, Instant((2021, 1, 1)), 1)).offset(1, DateUnit.YEAR)
+            >>> Period((DateUnit.YEAR, Instant((2021, 1, 1)), 1)).offset(
+            ...     1, DateUnit.YEAR
+            ... )
             Period((<DateUnit.YEAR: 'year'>, Instant((2022, 1, 1)), 1))
 
             >>> Period((DateUnit.DAY, Instant((2011, 2, 28)), 1)).offset(1)
@@ -722,15 +747,23 @@ class Period(tuple):
             Period((<DateUnit.YEAR: 'year'>, Instant((2014, 12, 31)), 1))
 
         """
+
+        start: None | t.Instant = self[1].offset(
+            offset, self[0] if unit is None else unit
+        )
+
+        if start is None:
+            raise NotImplementedError
+
         return self.__class__(
             (
                 self[0],
-                self[1].offset(offset, self[0] if unit is None else unit),
+                start,
                 self[2],
             ),
         )
 
-    def contains(self, other: Period) -> bool:
+    def contains(self, other: t.Period) -> bool:
         """Returns ``True`` if the period contains ``other``.
 
         For instance, ``period(2015)`` contains ``period(2015-01)``.
@@ -739,7 +772,7 @@ class Period(tuple):
         return self.start <= other.start and self.stop >= other.stop
 
     @property
-    def stop(self) -> Instant:
+    def stop(self) -> t.Instant:
         """Return the last day of the period as an Instant instance.
 
         Examples:
@@ -772,10 +805,9 @@ class Period(tuple):
 
         """
         unit, start_instant, size = self
-        year, month, day = start_instant
 
         if unit == DateUnit.ETERNITY:
-            return Instant((float("inf"), float("inf"), float("inf")))
+            return Instant.eternity()
 
         if unit == DateUnit.YEAR:
             date = start_instant.date.add(years=size, days=-1)
@@ -795,70 +827,92 @@ class Period(tuple):
 
         raise ValueError
 
+    @property
+    def is_eternal(self) -> bool:
+        return self == self.eternity()
+
     # Reference periods
 
     @property
-    def last_week(self) -> Period:
+    def last_week(self) -> t.Period:
         return self.first_week.offset(-1)
 
     @property
-    def last_fortnight(self) -> Period:
-        start: Instant = self.first_week.start
+    def last_fortnight(self) -> t.Period:
+        start: t.Instant = self.first_week.start
         return self.__class__((DateUnit.WEEK, start, 1)).offset(-2)
 
     @property
-    def last_2_weeks(self) -> Period:
-        start: Instant = self.first_week.start
+    def last_2_weeks(self) -> t.Period:
+        start: t.Instant = self.first_week.start
         return self.__class__((DateUnit.WEEK, start, 2)).offset(-2)
 
     @property
-    def last_26_weeks(self) -> Period:
-        start: Instant = self.first_week.start
+    def last_26_weeks(self) -> t.Period:
+        start: t.Instant = self.first_week.start
         return self.__class__((DateUnit.WEEK, start, 26)).offset(-26)
 
     @property
-    def last_52_weeks(self) -> Period:
-        start: Instant = self.first_week.start
+    def last_52_weeks(self) -> t.Period:
+        start: t.Instant = self.first_week.start
         return self.__class__((DateUnit.WEEK, start, 52)).offset(-52)
 
     @property
-    def last_month(self) -> Period:
+    def last_month(self) -> t.Period:
         return self.first_month.offset(-1)
 
     @property
-    def last_3_months(self) -> Period:
-        start: Instant = self.first_month.start
+    def last_3_months(self) -> t.Period:
+        start: t.Instant = self.first_month.start
         return self.__class__((DateUnit.MONTH, start, 3)).offset(-3)
 
     @property
-    def last_year(self) -> Period:
-        start: Instant = self.start.offset("first-of", DateUnit.YEAR)
+    def last_year(self) -> t.Period:
+        start: None | t.Instant = self.start.offset("first-of", DateUnit.YEAR)
+        if start is None:
+            raise NotImplementedError
         return self.__class__((DateUnit.YEAR, start, 1)).offset(-1)
 
     @property
-    def n_2(self) -> Period:
-        start: Instant = self.start.offset("first-of", DateUnit.YEAR)
+    def n_2(self) -> t.Period:
+        start: None | t.Instant = self.start.offset("first-of", DateUnit.YEAR)
+        if start is None:
+            raise NotImplementedError
         return self.__class__((DateUnit.YEAR, start, 1)).offset(-2)
 
     @property
-    def this_year(self) -> Period:
-        start: Instant = self.start.offset("first-of", DateUnit.YEAR)
+    def this_year(self) -> t.Period:
+        start: None | t.Instant = self.start.offset("first-of", DateUnit.YEAR)
+        if start is None:
+            raise NotImplementedError
         return self.__class__((DateUnit.YEAR, start, 1))
 
     @property
-    def first_month(self) -> Period:
-        start: Instant = self.start.offset("first-of", DateUnit.MONTH)
+    def first_month(self) -> t.Period:
+        start: None | t.Instant = self.start.offset("first-of", DateUnit.MONTH)
+        if start is None:
+            raise NotImplementedError
         return self.__class__((DateUnit.MONTH, start, 1))
 
     @property
-    def first_day(self) -> Period:
+    def first_day(self) -> t.Period:
         return self.__class__((DateUnit.DAY, self.start, 1))
 
     @property
-    def first_week(self) -> Period:
-        start: Instant = self.start.offset("first-of", DateUnit.WEEK)
+    def first_week(self) -> t.Period:
+        start: None | t.Instant = self.start.offset("first-of", DateUnit.WEEK)
+        if start is None:
+            raise NotImplementedError
         return self.__class__((DateUnit.WEEK, start, 1))
 
     @property
-    def first_weekday(self) -> Period:
+    def first_weekday(self) -> t.Period:
         return self.__class__((DateUnit.WEEKDAY, self.start, 1))
+
+    @classmethod
+    def eternity(cls) -> t.Period:
+        """Return an eternity period."""
+        return cls((DateUnit.ETERNITY, Instant.eternity(), -1))
+
+
+__all__ = ["Period"]