+++ openfisca_core/variables/variable.py
@@ -2,8 +2,6 @@ from __future__ import annotations
 
 from typing import NoReturn
 
-from openfisca_core.types import Formula, Instant
-
 import datetime
 import re
 import textwrap
@@ -11,7 +9,7 @@ import textwrap
 import numpy
 import sortedcontainers
 
-from openfisca_core import commons, periods
+from openfisca_core import commons, periods, types as t
 from openfisca_core.entities import Entity, GroupEntity
 from openfisca_core.indexed_enums import Enum, EnumArray
 from openfisca_core.periods import DateUnit, Period
@@ -385,8 +383,8 @@ class Variable:
 
     def get_formula(
         self,
-        period: Instant | Period | str | int = None,
-    ) -> Formula | None:
+        period: None | t.Instant | t.Period | str | int = None,
+    ) -> None | t.Formula:
         """Returns the formula to compute the variable at the given period.
 
         If no period is given and the variable has several formulas, the method
@@ -399,7 +397,7 @@ class Variable:
             Formula used to compute the variable.
 
         """
-        instant: Instant | None
+        instant: None | t.Instant
 
         if not self.formulas:
             return None